app dev archetype types binding and defaults
